CDC #322 - Adding user-defined fields data types to "shared" package; Updating DataTypes to re-export for backwards compatibility
